수정: 피드백 반영
입출력 쪽의 이름을 바꾸고, 입출력과 관계된 코드를 모두 view 쪽으로 옮겼다.
처음에는 안 좋지 않나 생각했지만, 일단 이대로 가볼 생각이다.
그나저나 정신없이 하다 보니 커밋 자주 하는 게 잘 안 된다.
