+++ src/main/java/lotto/model/LottoMaker.java
@@ -1,5 +1,7 @@
 package lotto.model;
 
+import java.util.List;
+
 public abstract class LottoMaker {
     LottoRule rule;
 
@@ -8,4 +10,6 @@ public abstract class LottoMaker {
     }
 
     public abstract Lotto getLotto();
+
+    public abstract List<Lotto> getAutoLottos(final int purchaseAmount);
 }
